+++ blocks/init/src/blocks/components/scroll-to-target/components/scroll-to-target-editor.js
@@ -1,0 +1,14 @@
+import { __ } from '@wordpress/i18n';
+import React from 'react'; // eslint-disable-line no-unused-vars
+
+export const ScrollToTargetEditor = (props) => {
+  const {
+    scrollText,
+  } = props;
+
+  return (
+    <button className="js-scroll-to-target scroll-to-target">
+      {scrollText}
+    </button>
+  );
+};